Crear interfaces CRUD y DTOs

+++ src/main/java/org/lht/m6preder/persistence/crud/CapacitacionCrudRepository.java
@@ -3,5 +3,6 @@ package org.lht.m6preder.persistence.crud;
 import org.lht.m6preder.persistence.entity.Capacitacion;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CapacitacionCrudRepository extends CrudRepository<Capacitacion, Integer> {
+
+public interface CapacitacionCrudRepository extends CrudRepository<Capacitacion, Long> {
 }
